requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
tavily-python
llama-parse
python-dotenv
nest_asyncio